Remove lint ignore
