+++ internal/driver/glfw/window.go
@@ -1535,17 +1535,3 @@ func isKeyModifier(keyName fyne.KeyName) bool {
 		keyName == desktop.KeyAltLeft || keyName == desktop.KeyAltRight ||
 		keyName == desktop.KeySuperLeft || keyName == desktop.KeySuperRight
 }
-
-//lint:ignore U1000 This is fine, just a helper function for some platforms
-func standardResize(w *window, size fyne.Size) {
-	d, ok := fyne.CurrentApp().Driver().(*gLDriver)
-	if !ok { // don't wait to redraw in this way if we are running on test
-		w.canvas.Resize(size)
-		return
-	}
-
-	runOnDraw(w, func() {
-		w.canvas.Resize(size)
-		d.repaintWindow(w)
-	})
-}
