+++ internal/driver/glfw/window_notwindows.go
@@ -1,0 +1,6 @@
+// +build !windows
+
+package glfw
+
+func (w *window) setDarkMode(dark bool) {
+}
