--- internal/driver/glfw/window_other.go
@@ -1,12 +1,0 @@
-// +build !linux,!windows
-
-package glfw
-
-import "fyne.io/fyne/v2"
-
-func (w *window) platformResize(canvasSize fyne.Size) {
-	standardResize(w, canvasSize)
-}
-
-func (w *window) setDarkMode(dark bool) {
-}
